meson: Don't disable the MSVC C4244 warnings
Stop disabling this warning, and fix the remaining cases in our
codebase:

    ../src/checkasm.c(771): warning C4244: '=': conversion from 'double' to 'uint64_t', possible loss of data
    ../src/stats.c(112): warning C4244: 'function': conversion from 'const uint64_t' to 'double', possible loss of data
    ../src/utils.c(246): warning C4244: '=': conversion from 'double' to 'float', possible loss of data

This avoids potential warnings in these files, if these files are
built in a foreign build system with different warning flag
configurations.

+++ src/checkasm.c
@@ -805,7 +805,7 @@ int checkasm_run(const CheckasmConfig *config)
             return 1;
         }
 
-        state.target_cycles = 1e3 * cfg.bench_usec / low_estimate;
+        state.target_cycles = (uint64_t) (1e3 * cfg.bench_usec / low_estimate);
         checkasm_measure_nop_cycles(&state.nop_cycles, state.target_cycles);
     }
